# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/check-email:
    post:
      operationId: checkEmail
      tags: [users]
      summary: Check if an email has saved data
      description: Public endpoint — checks whether a given email address is associated with an account that already has tasks, sprints, or custom columns. Used to decide whether to show a merge confirmation before sending a magic link.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckEmailRequest"
      responses:
        "200":
          description: Whether the email has existing data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckEmailResponse"
...
  /workspaces/{workspaceId}/tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: List tasks
      description: Returns all tasks for the given workspace.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      description: Creates a new task in the given workspace.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /workspaces/{workspaceId}/tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      description: Updates an existing task. May spawn a new recurrence task if the task has a recurrence rule and is being completed.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
        - $ref: "#/components/parameters/taskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Task updated, with optional spawned recurrence task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskUpdateResponse"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      description: Soft-deletes a task by setting its deletedAt timestamp.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
        - $ref: "#/components/parameters/taskId"
      responses:
        "204":
          description: Task deleted
  /workspaces/{workspaceId}/tasks/{taskId}/permanent:
    delete:
      operationId: deleteTaskPermanent
      tags: [tasks]
      summary: Permanently delete a task
      description: Hard-deletes a task, removing it from the database entirely. Intended for use from the Trash Bin "Delete forever" action. The task must already be soft-deleted (deletedAt set).
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
        - $ref: "#/components/parameters/taskId"
      responses:
        "204":
          description: Task permanently deleted
  /workspaces/{workspaceId}/tasks/bulk-archive:
    patch:
      operationId: bulkArchiveTasks
      tags: [tasks]
      summary: Bulk archive tasks
      description: Archives multiple tasks by setting their archivedAt timestamp.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkArchiveRequest"
      responses:
        "200":
          description: Tasks archived
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    workspaceId:
      name: workspaceId
      in: path
      required: true
      description: The workspace ID.
      schema:
        type: string
...
    taskId:
      name: taskId
      in: path
      required: true
      description: The task ID.
      schema:
        type: string
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        title:
          type: string
        notes:
          type: string
          default: ""
        columnId:
          type: string
        sprintId:
          type: string
          nullable: true
        storyPoints:
          type: integer
          nullable: true
        order:
          type: integer
        dueDate:
          type: string
          nullable: true
          description: YYYY-MM-DD
        inProgressAt:
          type: string
          nullable: true
          description: ISO datetime
        archivedAt:
          type: string
          nullable: true
          description: ISO datetime
        deletedAt:
          type: string
          nullable: true
          description: ISO datetime
        parentId:
          type: string
          nullable: true
        predecessorIds:
          type: array
          items:
            type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          nullable: true
        reminder:
          type: string
          nullable: true
        reminderDismissedAt:
          type: string
          nullable: true
        recurrence:
          type: string
          nullable: true
          enum:
            - daily
            - weekly
            - monthly
        createdAt:
          type: string
          description: ISO datetime
        updatedAt:
          type: string
          description: ISO datetime
      required:
        - id
        - workspaceId
        - title
        - notes
        - columnId
        - sprintId
        - storyPoints
        - order
        - dueDate
        - inProgressAt
        - archivedAt
        - deletedAt
        - parentId
        - predecessorIds
        - tags
        - reminder
        - reminderDismissedAt
        - recurrence
        - createdAt
        - updatedAt
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
        notes:
          type: string
          default: ""
        columnId:
          type: string
        sprintId:
          type: string
          nullable: true
        storyPoints:
          type: integer
          nullable: true
        order:
          type: integer
        dueDate:
          type: string
          nullable: true
          description: YYYY-MM-DD
        parentId:
          type: string
          nullable: true
        predecessorIds:
          type: array
          items:
            type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          nullable: true
        reminder:
          type: string
          nullable: true
        recurrence:
          type: string
          nullable: true
          enum:
            - daily
            - weekly
            - monthly
      required:
        - title
        - columnId
        - order
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        notes:
          type: string
        columnId:
          type: string
        sprintId:
          type: string
          nullable: true
        storyPoints:
          type: integer
          nullable: true
        order:
          type: integer
        dueDate:
          type: string
          nullable: true
          description: YYYY-MM-DD
        inProgressAt:
          type: string
          nullable: true
          description: ISO datetime
        archivedAt:
          type: string
          nullable: true
          description: ISO datetime
        deletedAt:
          type: string
          nullable: true
          description: ISO datetime
        parentId:
          type: string
          nullable: true
        predecessorIds:
          type: array
          items:
            type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          nullable: true
        reminder:
          type: string
          nullable: true
        reminderDismissedAt:
          type: string
          nullable: true
        recurrence:
          type: string
          nullable: true
          enum:
            - daily
            - weekly
            - monthly
    TaskUpdateResponse:
      type: object
      properties:
        updated:
          $ref: "#/components/schemas/Task"
        spawned:
          oneOf:
            - $ref: "#/components/schemas/Task"
            - type: "null"
      required:
        - updated
        - spawned
    BulkArchiveRequest:
      type: object
      properties:
        taskIds:
          type: array
          items:
            type: string
      required:
        - taskIds
...
    CheckEmailRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
    CheckEmailResponse:
      type: object
      required: [hasData]
      properties:
        hasData:
          type: boolean
...